requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.2
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cmake==3.27.7
confection==0.1.3
cymem==2.0.8
decorator==4.4.2
deep-translator==1.11.4
filelock==3.13.1
fsspec==2023.10.0
gTTS==2.4.0
huggingface-hub==0.17.3
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
langcodes==3.3.0
langdetect==1.0.9
lit==17.0.4
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numba==0.58.1
numpy==1.26.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper==20231106
packaging==23.2
Pillow==10.0.1
preshed==3.0.9
proglog==0.1.10
pydantic==2.4.2
pydantic_core==2.10.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.2.1
tiktoken==0.5.1
tokenizers==0.14.1
torch==2.0.1
tqdm==4.66.1
transformers==4.35.0
triton==2.0.0
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
wasabi==1.1.2
weasel==0.3.4
xx-ent-wiki-sm @ https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.7.0/xx_ent_wiki_sm-3.7.0-py3-none-any.whl
